+++ backend/src/DataFixtures/SystemVariablesFixtureLoader.php
@@ -25,7 +25,7 @@ class SystemVariablesFixtureLoader
             SystemVariable::ACCOUNT_DEACTIVATION_ACTIVE_DAYS    => ['30', SystemVariable::SCOPE_SETTINGS],
             SystemVariable::FETCH_STRUCTURE_NAME_ERROR_DAYS     => ['3=7,10=30', SystemVariable::SCOPE_SETTINGS],
             SystemVariable::ALLOW_CHARACTER_DELETION            => ['0',  SystemVariable::SCOPE_PUBLIC],
-            SystemVariable::ALLOW_LOGIN_MANAGED                 => ['0',  SystemVariable::SCOPE_SETTINGS],
+            SystemVariable::ALLOW_LOGIN_NO_SCOPES               => ['0',  SystemVariable::SCOPE_SETTINGS],
             SystemVariable::DISABLE_ALT_LOGIN                   => ['0',  SystemVariable::SCOPE_SETTINGS],
             SystemVariable::MAIL_INVALID_TOKEN_ACTIVE           => ['',   SystemVariable::SCOPE_SETTINGS],
             SystemVariable::MAIL_INVALID_TOKEN_ALLIANCES        => ['',   SystemVariable::SCOPE_SETTINGS],